Remove InputLoopService.
